aml_service/ci_cd/requirements.txt
azure-cli==2.1.0
azureml-sdk==1.0.69
